Remove Player Gibs and Death Cam on RoundStart
Created new component on gibs prefab that subscribes to GameStateChange event and deletes itself when a new round starts, also added same functionality to DeathCam component. This prevents the player from respawning when a new round starts but still having the DeathCam as primary cam as well as gibs still being active which hurts performance.